% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_passtimes.R
\name{get_passtimes}
\alias{get_passtimes}
\title{Gets ISS passtimes from API}
\usage{
get_passtimes(lat, long, elevation = 70, hours = 12)
}
\arguments{
\item{lat}{}
}
\value{
A data frame
}
\description{
Gets ISS passtimes from API
}
