% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_addition.R
\name{grouped_mean_tidy}
\alias{grouped_mean_tidy}
\title{Find means by group}
\usage{
grouped_mean_tidy(dat, col_1, col_2)
}
\arguments{
\item{dat}{A dataframe}

\item{col_1}{A unquoted column name}

\item{col_2}{A unquoted column name}
}
\value{
A summarized dataset with group means
}
\description{
Find means by group
}
