% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_addition.R
\name{simple_addition}
\alias{simple_addition}
\title{Add two column}
\usage{
simple_addition(dat, col_1, col_2)
}
\arguments{
\item{dat}{A dataframe}

\item{col_1}{A quoted column name}

\item{col_2}{A quoted column name}
}
\value{
A number which is the sum of the two inputs
}
\description{
Add two column
}
