% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_addition.R
\name{simple_addition_tidy}
\alias{simple_addition_tidy}
\title{Add two column}
\usage{
simple_addition_tidy(dat, col_1, col_2)
}
\arguments{
\item{dat}{A dataframe}

\item{col_1}{A unquoted column name}

\item{col_2}{A unquoted column name}
}
\value{
A dataset with new column summing the input cols
}
\description{
Add two column
}
